Add video download button.
Currently, if you want to download the video to the local device you have to wait until the video finishes loading then select the menu button, and then select Download. It can take a while for the video to finish loading while on mobile or if you are accessing Tube Archivist remotely on a slow connection. Having a button to immediately download the video to the local device is more convenient and gives the user a way to watch the video if their connection is slow by downloading it and watching it locally on the device.